Split jdbc 4.1 and jdbc 4.0 tests and run them according to the java version
