+++ org/postgresql/test/jdbc4/jdbc41/AbortTest.java
@@ -5,7 +5,7 @@
 *
 *-------------------------------------------------------------------------
 */
-package org.postgresql.test.jdbc4;
+package org.postgresql.test.jdbc4.jdbc41;
 
 import java.sql.*;
 import java.util.concurrent.Callable;
