+++ org/postgresql/test/jdbc4/jdbc41/CloseOnCompletionTest.java
@@ -1,4 +1,4 @@
-package org.postgresql.test.jdbc4;
+package org.postgresql.test.jdbc4.jdbc41;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
